requirements.txt
discord-together==1.1.1
py-cord[voice]==2.0.0b7
eyed3==0.9.6
youtube-dl==2021.2.22
youtube-search-python==1.4.7
validators==0.12.6
python-dotenv==0.17.1
tinydb==4.4.0
aiohttp-jinja2==1.5
aiohttp==3.7.4.post0
brotlipy==0.7.0
aiodns==3.0.0
aiofiles==0.4.0
aiosignal==1.2.0
charset-normalizer==2.0.7
cchardet==2.1.7
